add graph db interface
